requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
docopt==0.6.2
feedparser==6.0.2
idna==2.10
pdfkit==0.6.1
PyYAML==5.3.1
requests==2.25.1
sgmllib3k==1.0.0
soupsieve==2.1
urllib3==1.26.2
youtube-dl==2020.12.22
